8282729: Serial: Move BOT implementation to collector specific directory
