+++ src/hotspot/share/gc/serial/serialBlockOffsetTable.inline.hpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2000, 2019, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2000, 2022, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -22,10 +22,10 @@
  *
  */
 
-#ifndef SHARE_GC_SHARED_BLOCKOFFSETTABLE_INLINE_HPP
-#define SHARE_GC_SHARED_BLOCKOFFSETTABLE_INLINE_HPP
+#ifndef SHARE_GC_SERIAL_SERIALBLOCKOFFSETTABLE_INLINE_HPP
+#define SHARE_GC_SERIAL_SERIALBLOCKOFFSETTABLE_INLINE_HPP
 
-#include "gc/shared/blockOffsetTable.hpp"
+#include "gc/serial/serialBlockOffsetTable.hpp"
 
 #include "gc/shared/space.hpp"
 #include "runtime/safepoint.hpp"
@@ -69,4 +69,4 @@ inline void BlockOffsetSharedArray::check_reducing_assertion(bool reducing) {
            Thread::current()->is_ConcurrentGC_thread(), "Crack");
 }
 
-#endif // SHARE_GC_SHARED_BLOCKOFFSETTABLE_INLINE_HPP
+#endif // SHARE_GC_SERIAL_SERIALBLOCKOFFSETTABLE_INLINE_HPP
